Player now moves!!!
PlayerComponet is now PlayerInfoComponent. Makes more sense to me.
Velocity of player is changed through key presses. This concerns me
because in WorldSystem i use  setVelocity for  the body. What if player
was on moving platform and needed to move with it? it would slide from
underneath me. Thinking of using some addVelocity or whatever its called