Fixed problem with velocity in previous commit
Added new Vector2 to MovementComponent called previousVelocity. This way
i can manage how to subtract the right velocity and add the new one.
Check WorldSystem's processEntity() for implementation and explanation.
Maybe there is a better way of doing this, thought of having an array of
vectors for the MovementComponent, but it seemed to complex and
unintuitive.